requirements.txt
Flask==2.0.2
requests==2.26.0
python-dotenv==0.19.2
gunicorn==20.1.0
werkzeug==2.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
urllib3==1.26.20
certifi==2024.8.30
charset-normalizer==2.0.12
idna==3.10
setuptools==75.1.0
MarkupSafe==2.1.5